+++ packages/flame/lib/src/test_helpers/expect_vector2.dart
@@ -1,5 +1,6 @@
-import 'package:flame/extensions.dart';
 import 'package:test/test.dart';
+
+import '../../extensions.dart';
 
 void expectVector2(
   Vector2 actual,