+++ packages/flame/test/components/composed_component_test.dart
@@ -2,10 +2,8 @@ import 'dart:ui';
 
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
+import 'package:flame/test.dart';
 import 'package:test/test.dart';
-
-import '../util/mock_canvas.dart';
-import '../util/mock_gesture_events.dart';
 
 class MyGame extends BaseGame with HasTappableComponents {}