+++ packages/flame/test/extensions/canvas_test.dart
@@ -1,7 +1,6 @@
 import 'package:flame/extensions.dart';
+import 'package:flame/test.dart';
 import 'package:test/test.dart';
-
-import '../util/mock_canvas.dart';
 
 void main() {
   group('Canvas extensions tests', () {