+++ packages/flame/test/game/camera_and_viewport_test.dart
@@ -1,13 +1,10 @@
 import 'dart:ui' show Paint;
 
+import 'package:flame/components.dart';
 import 'package:flame/extensions.dart';
-import 'package:flame/src/anchor.dart';
-import 'package:flame/src/components/position_component.dart';
-import 'package:flame/src/game/base_game.dart';
-import 'package:flame/src/game/viewport.dart';
+import 'package:flame/game.dart';
+import 'package:flame/test.dart';
 import 'package:test/test.dart';
-
-import '../util/mock_canvas.dart';
 
 class TestComponent extends PositionComponent {
   static final Paint _paint = Paint();
@@ -127,7 +124,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0)', // camera translation
-          'translate(10.0, 10.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 10.0, 10.0, 0.0, 1.0)', // position component translate
           'translate(0.0, 0.0)', // position component anchor
         ],
       );
@@ -155,7 +152,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, -4.0, -4.0, 0.0, 1.0)', // camera translation
-          'translate(10.0, 10.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 10.0, 10.0, 0.0, 1.0)', // position component translate
           'translate(0.0, 0.0)', // position component anchor
         ],
       );
@@ -200,7 +197,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 40.0, 30.0, 0.0, 1.0)', // camera translation
-          'translate(10.0, 20.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 10.0, 20.0, 0.0, 1.0)', // position component translate
           'translate(-0.5, -0.5)', // position component anchor
         ],
         // result: 50 - w/2, 50 - h/2 (perfectly centered)
@@ -231,7 +228,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, -550.0, -1920.0, 0.0, 1.0)', // camera translation
-          'translate(600.0, 2000.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 600.0, 2000.0, 0.0, 1.0)', // position component translation
           'translate(-0.5, -0.5)', // position component anchor
         ],
         // result: 50 - w/2, 80 - h/2 (respects fractional relative offset)
@@ -323,7 +320,7 @@ void main() {
         ),
         [
           'transform(2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 1.0)', // camera translation and zoom
-          'translate(100.0, 100.0)', // position component
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 100.0, 100.0, 0.0, 1.0)', // position component translate
           'translate(-0.5, -0.5)', // anchor
         ],
       );
@@ -346,7 +343,7 @@ void main() {
         ),
         [
           'transform(2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 100.0, 100.0, 0.0, 1.0)', // camera translation and zoom
-          'translate(100.0, 100.0)', // position component
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 100.0, 100.0, 0.0, 1.0)', // position component translate
           'translate(-0.5, -0.5)', // anchor
         ],
       );