+++ packages/flame/CHANGELOG.md
@@ -22,6 +22,13 @@
  - Create `ColorEffect`
  - Adding ability to pause `SpriteAnimationComponent`
  - Adding `SpriteGroupComponent`
+ - Fix truncated last frame in non-looping animations
+ - Default size of `SpriteComponent` is `srcSize` instead of spritesheet size
+ - Export test helper methods
+ - Rename `ScaleEffect` to `SizeEffect`
+ - Introduce `scale` on `PositionComponent`
+ - Add `ScaleEffect` that works on `scale` instead of `size`
+ - Add class `NotifyingVector2`
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps